requirements.txt
fastapi
requests
python-dotenv
openai
psycopg2-binary
pgvector
uvicorn
langchain
langchain-openai
langchain-anthropic
langgraph
pydantic